@@ -1,7 +1,7 @@
 import { NavBar } from '../first_components/NavBar/NavBar';
 import { Header } from '../first_components/Header/Header';
 import { Footer } from '../first_components/Footer/Footer';
-import { RegionInfo } from '../secondary_components/RegionInfos/RegionInfos';
+import { InfoPage } from '../first_components/InfoPage/InfoPage';
 import { CardsRegions } from '../secondary_components/CardsRegions/CardsRegions';
 import { Separator } from '../first_components/Separator/Separator';
 
@@ -11,7 +11,7 @@ export const Regions = () => {
       <Header typeOfCarousel="deuxieme" />
       <NavBar />
       <Separator />
-      <RegionInfo />
+      <InfoPage />
       <Separator />
       <CardsRegions />
       <Footer />
